Feature/validation

+++ src/main/java/com/example/demoSpringboot/controller/CategoryController.java
@@ -1,7 +1,9 @@
 package com.example.demoSpringboot.controller;
 
 import com.example.demoSpringboot.model.Category;
+import com.example.demoSpringboot.model.Product;
 import com.example.demoSpringboot.service.CategoryService;
+import com.example.demoSpringboot.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,22 +24,13 @@ public class CategoryController {
     @Autowired
     CategoryService categoryService;
 
+    @Autowired
+    ProductService productService;
+
     @PostMapping
     public ResponseEntity<?> createNewCategory(@Valid @RequestBody Category category) {
         categoryService.save(category);
         return new ResponseEntity<>("Category valid", HttpStatus.CREATED);
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public Map<String, String> handleValidationException(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-        ex.getBindingResult().getAllErrors().forEach((error) -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = error.getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        return errors;
     }
 
     @GetMapping
@@ -52,9 +45,18 @@ public class CategoryController {
         return new ResponseEntity<>(categoryList, HttpStatus.OK);
     }
 
+    @GetMapping("/findCategoryByProductName")
+    public ResponseEntity<?> findCategoryByProductName(@RequestParam String name) {
+        List<Category> categoryList = categoryService.findCategoryByProductName(name);
+        return new ResponseEntity<>(categoryList, HttpStatus.OK);
+    }
+
     @GetMapping("{id}")
     public ResponseEntity<?> findCategoryById(@PathVariable Long id) {
         Optional<Category> categoryList = categoryService.findById(id);
+        if(!categoryList.isPresent()) {
+            return new ResponseEntity<>("Category does not exist", HttpStatus.NOT_FOUND);
+        }
         return new ResponseEntity<>(categoryList, HttpStatus.OK);
     }
 
@@ -68,6 +70,10 @@ public class CategoryController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<?> deleteCategory(@PathVariable Long id) {
+        List<Product> productList = productService.findProductByCategoryId(id);
+        if(!productList.isEmpty()) {
+            return new ResponseEntity<>("Category đang được sử dụng", HttpStatus.NO_CONTENT);
+        }
         categoryService.remove(id);
         return new ResponseEntity<>("Delete Completed", HttpStatus.OK);
     }
